opt: Split catalog.go into multiple files
Now that the Catalog types are in their own package, split the files
to separate the types better. This is a separate commit to make it
easier to review.

Release note: None
